+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ## Changed
 - Updated `Image` docs. (#270)
+- Update `libflate` dependency to `2.0.0`.
 - Fix some doc links. (#273)
 - Update ggez example to 0.9.3.