+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.11.2]
 ## Changed
 - Updated `Image` docs. (#270)
+- Update `libflate` dependency to `2.0.0`. (#279)
 - Fix some doc links. (#273)
 - Update ggez example to 0.9.3.